Build 15
Fixed issue wherein fnc_missionSpawner would hang if fnc_spawnGroup
returned a NULL-GROUP
This occured when the server had reached the max number of groups.
The fix was to periodically deleted empty groups (see fnc_mainThread).
this is probably a bug on Epoch only as Exile includes a check to delete
empty gorups in the server fsm.

Worked on the code to check for flags/plot poles but this stil needs
more  work.

worked on code to send notifications when a player trips an IED but this
needs more work.

Other minor bug-fixes