Update GMS_fnc_rewardKiller.sqf
Put in fix for respect loss bug on Exile servers. This is a temporary fix to deal with the fact that (vehicle _killer) and (objectParent _killer) seem to be broken, and because for some reason, the server can not read certain variables for the _killer object when the _killer is in a vehicle.